# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ActivityCalendarDay:
    properties:
      activity_count:
        description: Number of activities on this day
        type: integer
      activity_types:
        description: Types of activities done
        items:
          type: string
        type: array
      date:
        description: YYYY-MM-DD format
        type: string
      has_activity:
        description: Whether user was active on this day
        type: boolean
    type: object
  domain.ActivityCalendarResponse:
    properties:
      active_days:
        type: integer
      current_streak:
        type: integer
      longest_streak:
        type: integer
      summary:
        $ref: '#/definitions/domain.ActivityCalendarSummary'
      total_days:
        type: integer
      weeks:
        items:
          $ref: '#/definitions/domain.ActivityCalendarWeek'
        type: array
      year:
        type: integer
    type: object
  domain.ActivityCalendarSummary:
    properties:
      activity_breakdown:
        additionalProperties:
          type: integer
        description: Count by activity type
        type: object
      consecutive_weeks:
        description: Weeks with at least one activity
        type: integer
      most_active_count:
        description: Max activities in a single day
        type: integer
      most_active_day:
        description: Date with most activities
        type: string
      total_activities:
        type: integer
    type: object
  domain.ActivityCalendarWeek:
    properties:
      days:
        items:
          $ref: '#/definitions/domain.ActivityCalendarDay'
        type: array
    type: object
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.FreezeStreakRequest:
    properties:
      reason:
        example: Vacation
        type: string
    type: object
...
  domain.StreakInfo:
    properties:
      can_freeze:
        type: boolean
      current_streak:
        type: integer
      days_until_loss:
        type: integer
      freeze_count:
        type: integer
      last_activity_date:
        type: string
      longest_streak:
        type: integer
      max_freezes:
        type: integer
      streak_frozen:
        type: boolean
    type: object
  domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  entities.PracticeSentence:
    properties:
      id:
        type: string
      text:
        type: string
    type: object
  entities.PronunciationFeedback:
    properties:
      full_feedback_summary:
        type: string
      mispronouncedwords:
        items:
          type: string
        type: array
      overall_accuracy_score:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI-powered English coach for Ethiopians seeking global job opportunities
  title: LissanAI API
  version: "1.0"
paths:
...
  /api/v1/streak/activity:
    post:
      consumes:
      - application/json
      description: Record a user activity to maintain their streak - this is called
        internally by other services
      parameters:
      - description: Type of activity
        enum:
        - lesson_completed
        - quiz_passed
        - daily_goal_met
        in: query
        name: activity_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record activity (Internal)
      tags:
      - Streak
  /api/v1/streak/calendar:
    get:
      description: Get GitHub-like activity calendar showing daily learning activities
      parameters:
      - description: 'Year (default: current year)'
        example: 2025
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityCalendarResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get activity calendar
      tags:
      - Streak
  /api/v1/streak/freeze:
    post:
      consumes:
      - application/json
      description: Freeze the user's current streak to prevent it from being lost
        due to inactivity (limited uses per month)
      parameters:
      - description: Freeze reason
        in: body
        name: request
        schema:
          $ref: '#/definitions/domain.FreezeStreakRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Freeze streak
      tags:
      - Streak
  /api/v1/streak/info:
    get:
      description: Get the current user's streak information including current streak,
        longest streak, and freeze status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.StreakInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user streak information
      tags:
      - Streak
...
  /pronunciation/assess:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a target sentence and the user's recorded audio. It analyzes
        the user's speech against the target text and returns detailed feedback on
        their pronunciation. This is a multipart/form-data request.
      parameters:
      - description: The exact sentence the user was asked to say.
        in: formData
        name: target_text
        required: true
        type: string
      - description: The user's recorded audio file (e.g., in ogg, flac, or wav format).
        in: formData
        name: audio_data
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PronunciationFeedback'
        "400":
          description: Returns an error if the form data is invalid or missing.
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Returns an error if the AI service fails during assessment.
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Assess user's pronunciation
      tags:
      - Pronunciation
  /pronunciation/sentence:
    get:
      description: Dynamically generates a single, new English sentence tailored for
        pronunciation practice for Amharic speakers. Each request returns a unique
        sentence.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PracticeSentence'
        "500":
          description: Returns an error if the AI service fails to generate a sentence.
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a practice sentence
      tags:
      - Pronunciation
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"